remove duplicate code & rename table directories
