+++ frontend/components/tables/tutorials/edit-tutorials-table.tsx
